refactor: 모임 운동 캘린더 조회 서비스 분리
