+++ src/main/java/umc/cockple/demo/domain/exercise/controller/PartyExerciseController.java
@@ -6,7 +6,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RestController;
 import umc.cockple.demo.domain.exercise.controller.api.PartyExerciseApi;
 import umc.cockple.demo.domain.exercise.dto.PartyExerciseCalendarDTO;
-import umc.cockple.demo.domain.exercise.service.ExerciseQueryService;
+import umc.cockple.demo.domain.exercise.service.query.PartyExerciseQueryService;
 import umc.cockple.demo.global.response.BaseResponse;
 import umc.cockple.demo.global.response.code.status.CommonSuccessCode;
 import umc.cockple.demo.global.security.utils.SecurityUtil;
@@ -18,14 +18,14 @@ import java.time.LocalDate;
 @Validated
 public class PartyExerciseController implements PartyExerciseApi {
 
-    private final ExerciseQueryService exerciseQueryService;
+    private final PartyExerciseQueryService partyExerciseQueryService;
 
     @Override
     public ResponseEntity<BaseResponse<PartyExerciseCalendarDTO.Response>> getPartyExerciseCalender(
             Long partyId, LocalDate startDate, LocalDate endDate) {
         Long memberId = SecurityUtil.getCurrentMemberId();
 
-        PartyExerciseCalendarDTO.Response response = exerciseQueryService.getPartyExerciseCalendar(
+        PartyExerciseCalendarDTO.Response response = partyExerciseQueryService.getPartyExerciseCalendar(
                 partyId, memberId, startDate, endDate);
 
         return BaseResponse.of(CommonSuccessCode.OK, response);
